[tests] Refactor CubicBezierSpline
