+++ examples/CurveEditor/BezierUtils/CubicBezierApproximation.hpp
